backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RET v4 - Backend API with SQLite-based session caching"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.1",
    "argon2-cffi>=25.1.0",
    "cachetools>=5.3.0",
    "celery>=5.6.2",
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "gunicorn>=24.0.0",
    "jose>=1.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=0.1.0",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=0.0.1",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "lxml-stubs>=0.5.1",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pandas>=3.0.0",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "rarfile>=4.2",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.19.0",
    "svix>=1.84.1",
    "uvicorn>=0.40.0",
]
